+++ test/specs/options/title.spec.js
@@ -13,7 +13,7 @@ describe('title (option)', () => {
     expect(viewer.options.title).to.be.true;
   });
 
-  it('show not show title', (done) => {
+  it('should not show title', (done) => {
     const image = window.createImage();
     const viewer = new Viewer(image, {
       inline: true,
@@ -26,5 +26,39 @@ describe('title (option)', () => {
     });
 
     expect(viewer.options.title).to.be.false;
+  });
+
+  it('should support function', (done) => {
+    const image = window.createImage();
+    const viewer = new Viewer(image, {
+      inline: true,
+
+      title(img) {
+        return img.alt;
+      },
+
+      ready() {
+        expect(viewer.title.innerHTML).to.equal(image.alt);
+        done();
+      },
+    });
+
+    expect(viewer.options.title).to.be.a('function');
+  });
+
+  it('should support array', (done) => {
+    const image = window.createImage();
+    const viewer = new Viewer(image, {
+      inline: true,
+      title: [3, img => img.alt],
+
+      ready() {
+        expect(viewer.title.className).to.include('viewer-hide-sm-down');
+        expect(viewer.title.innerHTML).to.equal(image.alt);
+        done();
+      },
+    });
+
+    expect(viewer.options.title).to.be.an('array');
   });
 });